+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- RS-628: Support `ext` parameter in `Bucket.query`, [PR-125](https://github.com/reductstore/reduct-py/pull/125)
+
 ## [1.14.0] - 2025-02-25
 
 ### Added
 
-- RS-550: Add when condition to replication settings, [PR-123](https://github.com/reductstore/reduct-py/pull/123)
+- RS-550: Add `when` condition to replication settings, [PR-123](https://github.com/reductstore/reduct-py/pull/123)
 
 ## [1.13.0] - 2024-12-04